requirements.txt
autopep8==1.3.3
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
coverage==4.4.2
Django==1.11.7
django-silk==1.0.0
enum34==1.1.6
factory-boy==2.9.2
fake-factory==0.7.2
Faker==0.8.7
flake8==3.5.0
idna==2.6
ipaddress==1.0.17
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.13.3
pkg-resources==0.0.0
psycopg2==2.7.3.2
pycodestyle==2.3.0
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.10.0
sqlparse==0.2.4
text-unidecode==1.1
urllib3==1.22